+++ index.js
@@ -7,3 +7,4 @@
 export {MarkdownTransformer} from './src/markdown-transformer.js';
 export {HtmlTransformer} from './src/html-transformer.js';
 export {Transformer as JlReqTransformer} from './src/jlreq/transformer.js';
+export {NovelTransformer as PixivNovelTransformer} from './src/jlreq/pixiv/novel-transformer.js';