+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 [Unreleased]: https://github.com/deltanabla/markdown-transformer.js/compare/v0.0.0...HEAD
 
+### Added
+
+-   新しい変換機能を追加しました。
+    -   [pixiv小説記法](./src/jlreq/pixiv/novel-transformer.js)
+
 ## [0.0.0] - 2026-06-09
 
 [0.0.0]: https://github.com/deltanabla/markdown-transformer.js/releases/tag/v0.0.0